oops...

+++ app/Http/Controllers/Api/CampaignsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\v1;
 
 use App\Http\Requests;
 use App\Models\v1\Campaign;
@@ -8,7 +8,7 @@ use Dingo\Api\Contract\Http\Request;
 use App\Http\Requests\v1\CampaignRequest;
 use App\Http\Transformers\v1\CampaignTransformer;
 
-class CampaignsController extends ApiController
+class CampaignsController extends Controller
 {
 
     /**
@@ -35,6 +35,10 @@ class CampaignsController extends ApiController
      */
     public function index(Request $request)
     {
+        if ( ! $this->auth->user()->isAdmin()) {
+            abort(403);
+        }
+
         $campaigns = $this->campaign
                           ->filter($request->all())
                           ->paginate($request->get('per_page', 25));
@@ -50,6 +54,10 @@ class CampaignsController extends ApiController
      */
     public function show($id)
     {
+        if ( ! $this->auth->user()->isAdmin()) {
+            abort(403);
+        }
+
         $campaign = $this->campaign->findOrFail($id);
 
         return $this->response->item($campaign, new CampaignTransformer);
@@ -92,6 +100,10 @@ class CampaignsController extends ApiController
      */
     public function destroy($id)
     {
+        if ( ! $this->auth->user()->isAdmin()) {
+            abort(403);
+        }
+
         $campaign = $this->campaign->findOrFail($id);
 
         $campaign->delete();
